♻️ Refactor of tests cleaning up errors and speeding up testing
