+++ test/test_templates_blacklist.py
@@ -3,7 +3,7 @@ import os
 import unittest
 from os import getenv
 
-from test_templates import TestTemplates
+from base_templates import BaseTemplates
 
 from zabbixci import ZabbixCI
 from zabbixci.settings import Settings
@@ -15,7 +15,7 @@ DEV_ZABBIX_TOKEN = getenv("ZABBIX_TOKEN")
 DEV_GIT_REMOTE = getenv("REMOTE")
 
 
-class TestTemplatesBlacklist(TestTemplates):
+class TestTemplatesBlacklist(BaseTemplates, unittest.IsolatedAsyncioTestCase):
     def setUp(self):
         Settings.CACHE_PATH = "/tmp/zabbixci"
         self.cache = Cache(Settings.CACHE_PATH)
@@ -32,6 +32,7 @@ class TestTemplatesBlacklist(TestTemplates):
         Settings.ZABBIX_TOKEN = DEV_ZABBIX_TOKEN
         Settings.REMOTE = DEV_GIT_REMOTE
         Settings.SET_VERSION = True
+        Settings.REGEX_MATCHING = False
         Settings.TEMPLATE_BLACKLIST = "Not existing template,Not existing template 2"
 
         self.zci = ZabbixCI()
@@ -51,17 +52,6 @@ class TestTemplatesBlacklist(TestTemplates):
         self.assertFalse(changed, "Template deletion detected inside of blacklist")
 
         Settings.TEMPLATE_BLACKLIST = "Not existing template,Not existing template 2"
-
-        Settings.PULL_BRANCH = "test"
-
-        changed = await self.zci.pull()
-        self.assertTrue(changed, "Template was not restored")
-
-        Settings.PULL_BRANCH = "main"
-        changed = await self.zci.pull()
-        self.assertTrue(changed, "Template deletion from Git was not detected")
-
-        await self.restoreState()
 
 
 if __name__ == "__main__":
